Remove redundant GCM tests
The test vectors in gcm_tests.txt are just AES-GCM test vectors, all of
which are already present in our existing AES-GCM test vector files. To
test it we were just reimplementing AES-GCM with the low-level GCM bits.

With that, the only code remaining in gcm_test.cc was code we wrote,
added in 2019 in
https://boringssl-review.googlesource.com/c/boringssl/+/34266.

(Plus, with the exception of the last two, the test vectors, which used
to be embedded in the file, are actually just NIST test vectors. The
test driver code itself, since removed, does not appear to ever have
been derived from OpenSSL.)

This also removes the need to export the internal functions out of the
shared library build. (Though, in general, we probably should just be
turning those tests off in shared library builds.)

Change-Id: Ibb2b54f3c044a75a147f9fe8cc040b60e9a97482
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/73787
Reviewed-by: Adam Langley <agl@google.com>
Commit-Queue: Adam Langley <agl@google.com>
Auto-Submit: David Benjamin <davidben@google.com>