Add ML-DSA-87 in the FIPS module.
This is deliberately not exported in the public interface. Obviously we
could add it in the future if we wished but, for now, our suggestion is
to use ML-DSA-65. However, if needed in the future, it will already be
ready and FIPS validated inside the module.

Change-Id: I38a59d4e58b76cf74004432e87143380691a722b
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/73347
Reviewed-by: Bob Beck <bbe@google.com>
Commit-Queue: Adam Langley <agl@google.com>