Correctly re-ACK client Finished in DTLS 1.3
If client Finished gets through, but the server's responding ACK is
lost, the client will retransmit Finished (epoch 2). The server must
then process that retransmission and send another ACK.

However, the client may have since sent application data (epoch 3),
which the server may have since processed. Despite receiving data at
epoch 3, the server must keep epoch 2 open for some period of time to
accomodate this.

Unfortunately, there is no explicit signal in the protocol to do this.
The best we have is some guidance to be willing to do this for at least
2x MSL. See this discussion in the TLS mailing list:
https://mailarchive.ietf.org/arch/msg/tls/rof4SqkDrwU8o7WqSa9AO3YEUJI/

Implement this by keeping old epochs around for 2x MSL. This has the
side effect of also keeping old epochs around for a spell on KeyUpdate,
making us more resilient to packet reordering around KeyUpdate. For
simplicity, I've just get the same timeout for both, 2x MSL, or 4
minutes.

This opens a few cans of worms, because now record processing logic must
accomodate old epochs:

- The check against fragments from old epochs gets more complex.

- The check for maximum message size must move later; after the
  handshake, the server has a tight maximum message size, but it must be
  willing to receive (and skip over) retransmits from before the
  handshake when the limit was higher.

- Application data is fine. We already, at a lower layer, forbid any
  application data from coming in epochs 0 and 2.

- The ACK logic already did not make any particular assumptions here.
  Now we'll process ACKs from older epochs for slightly longer, but we
  already check that old epochs cannot ACK new epochs' messsages.

- ChangeCipherSpec is silly and would probably become moot once we start
  rejecting CCS in DTLS 1.3, but I just made it drop old epochs CCS on
  the floor to match the old behavior for now. (May as well avoid
  relying too much on DTLS 1.2 keeping one epoch around at once.)

This also requires tweaking the test slightly. flushHandshake needs to
run after the application data keys are installed, or the callback
cannot send data. That, in turn, means OutEpoch() is no longer the right
epoch for sending an ACK so some of the tests need to be updated to
still pick up epoch 2. (What's going on is that sending ACKs happens
before you construct the flight, but we want to test reordering, so we
simulate it afterwards.)

Bug: 42290594
Change-Id: I199b304fa9c3d8d252d258ab8ff231e3b5688e2e
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/74027
Auto-Submit: David Benjamin <davidben@google.com>
Reviewed-by: Nick Harper <nharper@chromium.org>
Commit-Queue: David Benjamin <davidben@google.com>