arithmetic: Remove `PartialEq` & `Debug` for  `LimbMask`.
Add `LimbMask::leak()` and change all callers to use it. This
proactively prevents accidental leakage of the `LimbMask` value
and makes it easier to audit the code for places where we
intentionally leak the value of a `LimbMask`.

Within the tests, use a `#[cfg(test)]-only wrapper `leak_in_test` to
make it easier to see that those leaks are uninteresting.