Fix pointer arithmetic in ASN1_item_ex_free
The ASN1_ITYPE_SEQUENCE loop does the 90s C thing where we didn't trust
our compilers and manually maintained loop variables as pointers. But
since this code is iterating backwards, it ends up creating a
one-before-the-start pointer. Unlike the one-past-the-end pointer, this
is not legal.

It never does anything with this pointer, and the compiler would need to
see through quite a few functions to observe this UB, so it is unlikely
that this UB had any impact in practice, but we should still fix it. Fix
this by writing this code more straightforwardly.

In doing so, push variable declarations closer to their use, so it's
easier to tell that i is an int.

Bug: 382414689
Change-Id: I35fb2d3aba563c5be86aaca29de31f339bfdc297
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/74089
Auto-Submit: David Benjamin <davidben@google.com>
Reviewed-by: Adam Langley <agl@google.com>
Commit-Queue: Adam Langley <agl@google.com>