Update copyright.pl to sync copyright lines from OpenSSL
OpenSSL's "copyright consolidation" script standardizes their various
old copyright headers on a new one. In doing so, it recomputed the
copyright year as follows:

- The end year was always 2016, when they ran the script.
- If the file began with an EAY copyright line, that starting year was
  used.
- Otherwise, the start year was ignored and recomputed from version
  control.

This final step will not run in BoringSSL, because we started a new
history. Instead, modify the script to simply take the result of the
process from the corresponding file in OpenSSL.

Bug: 364634028
Change-Id: I6083a398c7d742210d1b67110dda755ba0509f6c
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/74708
Commit-Queue: David Benjamin <davidben@google.com>
Reviewed-by: Adam Langley <agl@google.com>