aes: Clarify empty buffer handling in AES-CTR.
At one point we required `f` to be able to handle empty inputs,
but a later refactoring changed that, and added the necessary
check before calling `f`, but we forgot to update the comment.

To further clarify things, use `NonZeroU32` when incrementing
the counter, since a zero increment would result in a reused
counter.