Stop playing with time strings with strlcat
In ASN1_TIME_to_generalizedtime we were already parsing the
input time string to check for validity at the start of the
function. Instead of throwing away the tm we construct to do this,
keep the tm, and just convert the time normally to what we want
after the check instead of messing with the original string.

Change-Id: I32b351abd135b1f9c30b228c09718c5be994187a
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/74848
Reviewed-by: Adam Langley <agl@google.com>
Commit-Queue: Bob Beck <bbe@google.com>