digest/hmac: Clarify error handling.
Computing an HMAC can only fail if the input is too long. The
other `FinishError` condition, where we failed to meet the
precondition for `BlockContext::try_finish()`, is obviously,
using information local to the function, unreachable.

Similarly, when there is only a single slice of bytes as input,
then the `InputTooLongError` should be the `usize` type, not
`u64`. The case where the u64 -> usize condition fails is not
interesting, so we use a saturating conversion instead of
giving ourselves the future work of proving that it is
unreachable.

Finally, remove the `From<FinishError> for error::Unspecified`
implementation, in favor of `error::erase`, following the new
convention.