chacha: Clarify the dispatching to assembly implementations.
Clarify that there are 4 different functions, one for each architecture,
which have different capabilities w.r.t. overlapping input/output.

we plan to extend this to support all the details of the dispatching
that is currently within each assembly implementation of
`ChaCha20_ctr32` when that dispatching logic is moved to Rust.