arithmetic internals: Clarify memory safety of calls to `bn_mul_mont`.
Replace `debug_asesrt!`-based checking with proper error checking. The
error cases will never be reached because the callers already ensured
that the slices are the correct lengths, but this is more clearly
correct.

The previous step defining the `InOut` type didn't work out so well,
so replace `InOut` with `AliasingSlices` that does the same thing. The
cost is more monomorphization, but that will become moot soon, and it
already isn't too bad since there are only three cases to consider.
It does help reduce the number of length checks that end up getting
generated.