arithmetic internals: Restore old signature of bn_mul_mont.
In BoringSSL, it returns `int`. Have it do that again, to make the
next bit of refactoring easier, where we replace `bn_mul_mont` with a
Rust function that dispatches to a bunch of other functions, all of
which also are declared as having the same semantics.

The good thing is that this restores the assembly code to match
BoringSSL.

The bad thing is that this some bits of dead code to handle zero
return values that we'll never see.