arithmetic/limbs internals: Clarify limbs < limbs checks.
Reject empty or non-matching-length inputs. Refactor the internal
API so that it is clearer that we're nearly always using this
function to reject too-large inputs. This makes it easier to see
that it is OK to return an error when previously we didn't.

This isn't intended to have any effect other than making the code
easier to audit.