build.rs: Disable LTO if enabled implictly
LTO between Rust and C is complicated. It requires clang and specific
invocations configurations for both Rust and C compilations.
If the CFLAGS from the environment enable LTO, for example from
distribution build tools this can lead to errors.
Instead always disable LTO for the C compilation. There is not much to
optimize away anyways.

Closes #1444