arithmetic: Use `AsChunks` to clarify `sqr_mont5`.
Replace the use of `ControlFlow` with stronger input types.

Besides `ControlFlow` being an unusual thing for this codebase, this
also is a step towards making `bn_sqr8x_mont` easier to call from
other callers in the future.

Rename `bn_sqr8x_mont` to `sqr_mont5` to match the naming of the other
`...5` functions in the module.