bigint: Move some memory allocations out of low-level functions.
Make allocating and zeroing memory for `Elem`s more explicit, towards
the goal of eventually removing allocations from low-level functions.

Whenever a function returns an `Elem` by value, have it accept a
`Storage` that will be recycled into the return value. Effectively,
this moves the allocation from the callee to the caller.