util/fipstools: generalize hkdf KDA subprocess
Previously the only KDA ACVP algorithm mode supported by the acvptool
was HKDF mode, and so the file was called hkdf.go and didn't interrogate
the vector set mode.

In preparation for supporting the OneStepNoCounter KDA mode we need to
generalize the subprocess handler as both it and HKDF KDA are advertised
as "KDA" algorithms and must be multiplexed on the "mode".

To simplify review this commit moves the existing HKDF code, and adds
the small bit of multi-plexing required to dispatch by mode, but doesn't
introduce any other changes.

Change-Id: I3ad640d6b3fa93483c8fac6cadcfb88cf9da5430
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/75927
Reviewed-by: Adam Langley <agl@google.com>
Reviewed-by: Bob Beck <bbe@google.com>
Commit-Queue: Bob Beck <bbe@google.com>