aead::chacha internals: Stop checking FXSR bit.
The checking for the FXSR bit is an anachronism from OpenSSL, where
it was used as a way to indicate that SSE instructions should not be
used. Currently we require SSE/SSE2 to be available. In the future,
if we make SSE/SSE2 optional, then we'll need to condition the SSSE3
feature on SSE (operating system) support.

Remove the `Fxsr` CPU feature as we no longer check for it.