have old binutils support on avx10 as well and improve it
Currently there are no ymm/zmm in the avx10 code, but there will be.

I tested that when you fill out the map correctly, this code will work
for AVX-512 as well, being able to result in the same objdump if you
use make-avx-map-for-old-binutils to generate an asmMap.

Made the perl code less sensitive to whitespace differences, I still
think it's pretty understandable.

This currently does not change anything, not even the .S output.