arithmetic: Disambiguate meaning of two-argument `AliasingSlices3`.
Is `(r, a)` supposed to be interpreted as `(r, r, a)` or `(r, a, a)`?
When I created this, I assumed it was obvious that it should be
`(r, r, a)`. But, soon we'll need to be able to support the other
interpretation. Instead, have the caller write `(InOut(r), a)` to
make it obvious that `(r, r, a)`. is intended.