arithmetic/aarch64: Use optimized squaring in `elem_exp_consttime`.
Fix a performance regression from when we replaced `bn_mul_mont`.

`bn_mul_mont` used to check whether its two source operand pointers
were equal, calling the squaring function if so. Our expectation
when it was replaced was that every caller would statically know
whether it was doing squaring or non-squaring, so this check was
not needed. Unfortunately, we overlooked this one case.

The new `inout` machinery here allows converting an
`(r, a): AliasingSlices2<T>` to an `AliasingSlices3<T>` where the
source (`a`) is aliased instead of the destination ('r').