arithmetic: Simplify calculation of bit length.
Replace `limb::limbs_minimal_bits` with `limb::limb_minimal_bits` so
that we only ever look at the value of the highest bit. This is a no-
op change, as previously the outer loop in `limbs_minimal_bits` was
never executed more than once. In the sole caller, document why we
only have to look at the high limb.