digest/hmac: Make `Context` types a bit smaller.
Reduce the size of each `Context` type by 8 bytes.

Also, eliminate a redundant store of `num_pending` that was found as
a side effect.

There was a comment about aligning the pending buffer. However, it
seems like aligning it isn't important, especially as we didn't
bother to do so in the HMAC implementation. Probably it was a (likely
misguided) optimization idea. So just remove the comment without doing
anything to force an alignment.