aes_gcm/x86_64: More precisely model CLMUL-AVX-MOVBE key type.
The comments that `Xi` and `HTable` "should" be aligned is confusing.
It might have mattered for performance on some old CPUs but it's no
longer important, so remove them.

The comments also erroneously claimed that only 9 entries were needed,
due to me misreading the assembly code when I wrote these comments.
In fact, 12 entries are required.