aes: Better model `AES_KEY` and avoid unnecssary zeroization.
It is really a union, though C implementations don't consider it
such for C/C++-specific reasons. It is better for us to represent
it as a union in Rust so that the partial initialization of 128-bit
keys is communicated to the tooling. This way, we can avoid zeroing
the key before initializing it.