aes internals: Use `debug_assert_eq!()` for always-true condition.
With `assert_eq!()`, the compiler will not use NRVO, so the key will
first be constructed on the stack and then copied. This assertion
isn't really helpful as we've already ensured that the failure case
of all these C initialization functions will never be triggered by
the structure of `KeyValue`, so switching to `debug_assert_eq!` is
safe.