aes/aarch64: Simplify single-block encryption.
We overlooked that ctr32_encrypt_blocks is optimized for even numbers
of blocks; it does redundant work for a single block. Since we only
use it for single-block encryption, this was a waste. Fix this by
basing the single-block encryption on the old aes_hw_encrypt_block,
modified to do the XOR. The new version takes separate `rd_key` and
`rounds` parameters since we don't use `AES_KEY` on this target.