rsa: Use `PublicKeyComponents` consistently in verification.
We did some redundant validation that inputs were positive during
parsing of an RSA public key. Constructing a `PublicModulus` with a
zero value or a `PublicExponent` with a zero value will fail
correctly, so there is no need for the redundant check, except
perhaps to avoid allocating memory for invalid values. However, we
should solve the validate-before-allocating problem more generally.