bigint/x86_64: Eliminate most unnecessary chunking.
We don't actually need to pass arguments as `&[[Limb; 8]]` instead of
`&[Limb]` in most cases. As long as one argument has a type like that,
then we've ensured that the length is a multiple of 512 bits. In most
cases we were calling `as_chunks[_mut]` on something and then
immediately calling `as_flattened[_mut]` on it.